requirements.txt
requests==2.25.1
XlsxWriter==1.3.7
python-dateutil==2.8.1